replaced suppressed errors with error elimination
This now checks for the existence of the files either in the current
path or the include_path before loading.

In addition, mink/autoload.php loads 'vendor/.composer/autoload.php'
and if mink and if it exits then it doesn't need to load again.

Not tested beyond 'Works for me!'