Update Db.php
This array grows if the inserted id can't be removed by this function resulting in an ever growing list of "could\'t delete record $X from $tablename" throughout a test run. The full list is output at the end of each test within a single run. An alternative would be to clean up this array between test runs or setting ```$this->insertedIds = array();``` at the end of this function.