Updates for WebDriver documentation.
I added a "Locating Elements" section to the docstring for WebDriver.php with details on how
locators are specified and processed. That was a stumbling point for me when I first learned how to
use this module. I introduced the term "fuzzy locator" to refer to plain-string locators.

I reworded a lot of the other method docstrings to read better and correct grammatical errors. A few
methods (e.g. seeElement) duplicated the docstring for the corresponding interface method, so I
removed those.