Improvement to the unit test cases examples
The unit test examples that use database functionality like the seeInDatabase() method call need to have the Db module listed in the unit suite configuration file. The module was not enabled by default after running bootstrap and generate:test, so I think it would be very helpful to mention this detail on the guide for Unit Tests. It took me a while to figure out why the test run failed with an exception about the missing function in the UnitTester class.

Thank you, Markus