Only cleanup specified tables - MySQL
Config can now contain a clean_tables sequence of tables that will be
cleaned during cleanup(). If this list is missing/empty then all tables
will be cleaned. If it is not missing/empty then only tables in this
sequence will be cleaned up. Other tables will be completely ignored
unless also included in a provided dump. This first implementation is ONLY
for MySQL