Fix "Array to string conversion" errors
The "fail" method of both Codeception\PHPUnit\Constraint\WebDriver and WebDriverNot takes in a
$selector for error reporting purposes. The $selector is assumed to be a string and interpolated
into the error message, but there are cases where the $selector is actually an array or WebDriverBy
object. When that's the case, this method triggers a "Array to string conversion" PHP notice.

I noticed this in a test case I have that calls Codeception\Module\WebDriver->see():
    $I->see('Reviewers', ["css" => ".submission_nav .box.current"])

This test failed because the ".submission_nav" element disappeared, but instead of a descriptive
failure message I got "[PHPUnit_Framework_Exception] Array to string conversion".

To reduce code duplication, I extracted the logic used in ElementNotFound for converting a
selector into a string and put it in a new method Codeception\Util\Locator->humanReadableString().