Update ZF2 module
Update the doRequest() method in the ZF2 module's connector to grab the
application's response *after* the application has run. Before, the method
grabbed the response *before* the application was run. This resulted in an
empty response when other ZF libraries like Api Agility replaced the
application's response object with their own.