Use Laravel UploadedFile class for requests
Using Symfony’s UploadedFile has the effect that when Laravel’s Request class does convertUploadedFiles, all files are converted to Laravel’s UploadedFile class, but with the test flag set to false.

This, in turn, makes it difficult to use the class in a test, because the class then must pass the is_uploaded_file test in UploadedFile::isValid().