Some improvements
Two questions:

* > The methods of the $I object are taken from the PhpBrowser module. We will briefly describe them here:  
Don't you think that it would be better to show just 1-2 examples here (and refer to the module for the rest)? With the PhpBrowser section being so long, the WebDriver section gets moved very far down.

* > Important notice: PhantomJS is not maintained anymore. Use it at your own risk.  
Where is this text coming from? Can't see it in the source code.... I wanted to add the Headless Chrome link inside this message box.