Update appveyor.yml to specify php version 7.1
Pull request #4799 fails due to AppVeyor and Chocolatey installing PHP 7.2 and not constrained to any version. This entire file assumes PHP 7.1, so I added what I believe is the constraint for Chocolatey/cist. 

Disclaimer: I am not familiar with Chocolatey or AppVeyor so I'm kind of taking a stab in the dark, well an educated guess have you.