Clarify `grabFromDatabase` and `grabColumnFromDatabase`
The methods does not fetching "all values", but a single column.

It seems that `grabFromDatabase` and `grabColumnFromDatabase` doing the same:
I recomend to keep `grabFromDatabase` method but rename it to
`grabColumnFromDatabase` because it's more verbose.

I'm not sure that it's accepting comparison expressions or not, but if it
isn't, then the documentation about this should be affected in another PR (see
src/Codeception/Lib/Interfaces/Db.php:67).